+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 1.2.0
- - [CHANGED] Optional field can now be hidden (#2)
+ - [ADDED] Support for Bootstrap3. (#1)
+ - [CHANGED] Optional field can now be hidden. (#2)
  - [CHANGED] Card title is now configurable. (#5)
  - [CHANGED] Headers in Trello are now configurable. (#4)
  - [CHANGED] Reload URL is now configurable. (#3)